[style] integration / delete_file()をcommonに移動
