+++ test/webserv/integration/test_http_post.py
@@ -1,8 +1,8 @@
 import os
-import time
 
 import pytest
-from common_functions import read_file, send_request_and_assert_response
+from common_functions import (delete_file, read_file,
+                              send_request_and_assert_response)
 from common_response import (bad_request_response, created_response_close,
                              created_response_keep, no_content_response_close,
                              not_allowed_response, payload_too_large_response,
@@ -16,20 +16,6 @@ REQUEST_POST_4XX_DIR = REQUEST_DIR + "post/4xx/"
 UPLOAD_DIR = "root/upload/"
 UPLOAD_FILE_PATH = UPLOAD_DIR + "test_upload_file"
 CHUNKED_FILE_PATH = UPLOAD_DIR + "chunked_request_file"
-
-
-def delete_file(file_path):
-    if file_path is None:
-        return
-
-    try:
-        if os.path.exists(file_path):
-            os.remove(file_path)
-            time.sleep(1)
-            print(f"Deleted file: {file_path}")
-    except Exception as e:
-        print(f"Error deleting file: {file_path}, {e}")
-        raise AssertionError
 
 
 def assert_uploaded_file_content(upload_file_path, expected_upload_file_content):
